修改文件名的命名，解决裁剪方式上传图片失败的bug
